refactor: changed folder structure for  webapi configuration
