+++ src/WebApi/Config/SerilogSetup/ColumWriters/BoolColumnWriter.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.WebApi.Config;
+﻿namespace IAVH.BioTablero.CM.WebApi.Config.SerilogSetup.ColumWriters;
 
 using System;
