+++ src/WebApi/Config/SerilogSetup/Enrichers/IdEnricher.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.WebApi.Config;
+﻿namespace IAVH.BioTablero.CM.WebApi.Config.SerilogSetup.Enrichers;
 
 using System;
