+++ src/WebApi/Config/SwaggerSetup/Filters/CustomODataQueryOptions.cs
@@ -1,4 +1,4 @@
-namespace IAVH.BioTablero.CM.WebApi.Config;
+namespace IAVH.BioTablero.CM.WebApi.Config.SwaggerSetup.Filters;
 
 using System.Linq;
